refactor: split services into dedicated modules
