+++ domain/services/__init__.py
@@ -1,0 +1,17 @@
+from .trader import Trader
+from .symbol_registry import SymbolRegistry
+from .rest_client import RestClient
+from .ws_client import WsClient
+from .signal_engine import SignalEngine
+from .risk_manager import RiskManager
+from .trade_manager import TradeManager
+
+__all__ = [
+    "Trader",
+    "SymbolRegistry",
+    "RestClient",
+    "WsClient",
+    "SignalEngine",
+    "RiskManager",
+    "TradeManager",
+]
